requirements.txt
aioredis==2.0.1
amqp==5.1.1
anyio==3.6.2
arabic-reshaper==2.1.4
asgiref==3.6.0
asn1crypto==1.5.1
astroid==2.12.13
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
autopep8==2.0.1
billiard==3.6.4.0
black==22.12.0
bleach==5.0.1
Brotli==1.0.9
cachetools==5.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
cssselect2==0.7.0
daphne==4.0.0
data-wizard==2.0.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dill==0.3.6
dj-rest-auth==2.2.5
Django==4.1.4
django-admin-sortable2==2.1.4
django-allauth==0.51.0
django-anymail==9.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-crum==0.7.9
django-etc==1.4.0
django-extensions==3.2.1
django-filter==22.1
django-fsm==2.8.1
django-fsm-admin==1.2.5
django-import-export==3.0.2
django-jazzmin==2.6.0
django-jet-reboot==1.3.3
django-js-asset==2.0.0
django-json-widget==1.1.1
django-map-widgets==0.4.1
django-model-utils==4.3.1
django-oauth-toolkit==2.2.0
django-object-actions==4.1.0
django-pgconnection==1.0.3
django-pgtrigger==4.6.0
django-phonenumber-field==7.0.1
django-push-notifications==3.0.0
django-renderpdf==3.0.1
django-solo==2.0.0
django-storages==1.13.1
django-timezone-field==5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
docutils==0.19
drf-yasg==1.21.4
drfpasswordless==1.5.8
ecdsa==0.18.0
et-xmlfile==1.1.0
feedparser==6.0.10
fonttools==4.38.0
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
graphviz==0.20.1
gunicorn==20.1.0
h11==0.14.0
hiredis==2.1.0
html-json-forms==1.1.1
html5lib==1.1
http-ece==1.1.0
httplib2==0.21.0
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
isort==5.11.4
itertable==2.1.0
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
jsonschema==4.17.3
jwcrypto==1.4.2
kombu==5.2.4
lazy-object-proxy==1.8.0
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.4
mypy-extensions==0.4.3
natural-keys==2.0.0
num2words==0.5.12
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
packaging==22.0
pathspec==0.10.3
phonenumberslite==8.13.3
Pillow==9.3.0
pip-review==1.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psycopg2==2.9.5
py-vapid==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pydot==1.4.2
pydyf==0.5.0
pygraphviz==1.10
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyphen==0.13.2
pyrsistent==0.19.2
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywebpush==1.14.0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.3.0
redis==4.4.0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.12.1
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
svglib==1.4.1
swapper==1.3.0
tablib==3.3.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
twilio==7.16.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.5
weasyprint==57.1
webencodings==0.5.1
websockets==10.4
whitenoise==6.2.0
wrapt==1.14.1
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.5.2
zopfli==0.2.2
